refactor(gift-ranking): 실시간 급상승 랭킹 Suspense, ErrorBoundary, 변경된 훅을 사용하도록 리팩터링
- Suspense, ErrorBoundary 사용을 위해 랭킹을 나타내는 부분을 따로 컴포넌트로 분리
